requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pinecone-client>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
aiofiles>=23.2.0
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn
matplotlib
seaborn
plotly
openpyxl
xlsxwriter
reportlab
PyPDF2
pdfplumber
nltk
spacy
transformers
torch
sentence-transformers
chromadb
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
psutil
tqdm
colorama
rich
typer
click
pydantic
sqlalchemy
alembic
redis
celery
pytest
pytest-asyncio
pytest-cov
black
flake8
mypy
pre-commit
jupyter
ipykernel
notebook
jupyterlab